DOCUMENT: add comments and rename scripts
