--- Assets/Scripts/Enemy/EnemyNavigation2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1f6222cac63511b44a26377eab28587d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
